ci: Explicitly use Debian Bullseye-based Go image
It looks like the maintainers of the Go Docker image decided to switch
to base their images off of Debian Bookworm instead of Bullseye. This
meant that `golang-1.20` pulled an image with a newer version of glibc
which only broke when trying to run the image.

This is a quick fix that explicitly calls for the Bullseye version of
the build image.

Closes #22, #16.