Refactor endpoint handling and reconcilliation (#21)
* Disable cgo

As part of the Go 1.20 release it seems like the default for `CGO_ENABLED` is no longer carried over from the tools. This leads to linking issues on systems that use different versions of glibc from what the base image uses. See https://github.com/golang/go/issues/58550 for more details.

This change should fix #16

* Slim down the final image.

Use `scratch` as a base image since we're generating a static binary
anyway. Also be more explicity about the platform and target OS during
the build.

* Refactor endpoint handling and reconcilliation.

DRAFT

With the release of Nomad 1.6 it's possible to get the network address
of the allocation from Nomad. The change to enable this is only in the
client library and does not require updating the Nomad server. The IP
was sent back by older Nomad versions, it just wasn't available in the
client.

This enables refactoring the endpoint reconcilliation to make use of the
IP address to identify the endpoint within Cilium.

There is no longer a dependency on Consul for policies.

Additional, endpoints are now labelled with the task group and task
information as services can be created at those levels.

* Tidy

* Update the readme and add some basic tests.

Remove unused flags from the readme and command line and refactor the
code to allow for testing.

* Slightly better logging of labels

* Always update labels on reconcile.

* Fixes and align with upstream Dockerfile

* Missed one

* Reduce Dockerfile diff further

* Restore periodic reconcilliation of endpoints

* Remove stray return