Filter processed allocation events more.
The allocation events from Nomad include the NodeID where the allocation
is assigned. We can use that to drop events early in processing by
comparing the event NodeID with the local agent NodeID.

This reduces the load on Cilium agents in the cluster as they are no
longer queried about allocations that don't exist on the node.

Additionally we only need to watch the `Allocation` event stream and can
ignore events for terminal allocations. Nomad will handle cleaning up
the endpoint in that case.