Use Cilium kvstore and support multiple policies.
This change modifies the code base to use the Cilium kvstore package to
manage netreap policy data and its interactions with Cilium. This allows
netreap to match the configuration of the Cilium agent on the host and
enables the introduction of multiple policies. Configuation can still
be overridden but default will be requires from the Cilium agent.

Multiple policies can now also be managed under a prefix. Any values
within the prefix will be treated as Cilium policy documents and passed
to the local agent.

Note that this is a breaking change as the location of policy documents
in the kvstore has changed from `netreap.io/policy` to `netreap/policies/v1`
to match the patterns of Cilium.

Also this change currently depends on Cilium 1.15 as it uses API parameters
of the `PutPolicy` call that were unavailable in earlier versions.