Exit on Nomad event stream error.
We've found that it's better to exit immediately on any error from the Nomad event stream. Errors can occur any time but typically during a rolling update of the Nomad servers. Attempting to ignore some error types results in an infinite loop of

```
2024-03-06T04:11:05.089Z	WARN	reapers/endpoints.go:89	Ignoring invalid events payload from Nomad	{"error": "invalid character 's' looking for beginning of value"}
github.com/cosmonic-labs/netreap/reapers.(*EndpointReaper).Run.func1
	/app/netreap/reapers/endpoints.go:89
```

The Nomad job definition for netreap should result in a restart of the daemon process.